backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
sqlalchemy==2.0.25
psycopg[binary]==3.2.13
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
cryptography==41.0.7
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0
